refactor:  update & delete event endpoint using slug as the identifier
